added restful api call funcs to utils.py and currently trying to change the resolver function for RA and dec as a proof of concept
